add: Calc distance logic. View toolbar. Center point.

+++ app/src/main/java/by/godevelopment/fifthtask/domain/GeographicPointModel.kt
@@ -4,7 +4,7 @@ data class GeographicPointModel(
     val id: Int,
     val tittle_type: String? = null,
     val snippet_address: String? = null,
-    val distance: Double? = null,
+    val distance: Float? = null,
     val latitude: Double? = null,
     val longitude: Double? = null
 )
